Migrate unique Heatmap controls (#9360)
* Migrate Heatmap controls

* Lint
